Force install of local wheel when testing
pytest was failing because the pip install is downloading the latest published
version of the module instead of using the locally-built one. The docs say that
there is no ordering in the locations that are searched, so we can't say to
prefer the local wheel.

Disabling the index when installing `cedarpy` appears to solve the issue.